waze-e-meteorologia/requirements.txt
pandas
requests
python-dateutil
pytz
tenacity
geopy